# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: GenBI API
  version: 0.1.0
  description: Generative BI Platform API
servers:
  - url: http://localhost:8000/api/v1
    description: Development
paths:
...
  /admin/stats:
    get:
      operationId: admin_platform_stats
      summary: "(Platform counters — superuser only"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Platform stats
        "403":
          description: NOT_PLATFORM_ADMIN
  /admin/tenants:
    get:
      operationId: admin_list_tenants
      summary: "(Tenant list with counters — superuser only"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Tenant list
    post:
      operationId: admin_provision_tenant
      summary: "(Provision tenant + initial admin user (transactional)"
      security:
        - BearerAuth: []
      responses:
        "201":
          description: Tenant created
        "409":
          description: TENANT_EXISTS / USER_EXISTS
  /admin/tenants/{tenant_id}:
    get:
      operationId: admin_get_tenant
      summary: "(Tenant detail (users, schedules, counts, audit)"
      security:
        - BearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tenant detail
        "404":
          description: TENANT_NOT_FOUND
    patch:
      operationId: admin_update_tenant
      summary: "(Rename / suspend / activate / settings"
      security:
        - BearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Updated tenant
    delete:
      operationId: admin_delete_tenant
      summary: "(Decommission — guarded, refuses non-empty tenants without force"
      security:
        - BearerAuth: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: confirm
          in: query
          required: true
          schema:
            type: string
            enum: ["yes"]
        - name: force
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Deleted
        "422":
          description: TENANT_NOT_EMPTY
  /admin/admins:
    get:
      operationId: admin_list_superusers
      summary: "(Active superusers + grant history"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Superuser list
    post:
      operationId: admin_grant_superuser
      summary: "(Grant platform-superuser (by user id or email)"
      security:
        - BearerAuth: []
      responses:
        "201":
          description: Granted
    delete:
      operationId: admin_revoke_superuser
      summary: "(Revoke — effective within ~60s (cached re-check)"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Revoked
  /admin/audit:
    get:
      operationId: admin_audit_feed
      summary: "(Admin-action feed (actor/target/tenant filters)"
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Audit entries
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
...